Start separating out tests
