+++ tests/unit/alldeps/test_public_interface.py
@@ -1,0 +1,1 @@
+from records_mover.airflow.hooks import records_hook  # noqa
