+++ tests/unit/test_public_interface.py
@@ -1,4 +1,3 @@
-from records_mover.airflow.hooks import records_hook  # noqa
 from records_mover import Session  # noqa
 from records_mover.records import ExistingTableHandling  # noqa
 from records_mover.records import DelimitedRecordsFormat  # noqa
